build: split .c files into .c and .h
