+++ bsdiff.c
@@ -25,24 +25,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <stddef.h>
-#include <stdint.h>
-
-struct bsdiff_stream
-{
-	void* opaque;
-
-	void* (*malloc)(size_t size);
-	void (*free)(void* ptr);
-	int (*write)(struct bsdiff_stream* stream, const void* buffer, int size);
-};
-
-int bsdiff(const uint8_t* old, int64_t oldsize, const uint8_t* new, int64_t newsize, struct bsdiff_stream* stream);
-
-#if !defined(BSDIFF_HEADER_ONLY)
-
-#include <limits.h>
-#include <string.h>
+#include "bsdiff.h"
 
 #define MIN(x,y) (((x)<(y)) ? (x) : (y))
 
@@ -455,7 +438,5 @@ int main(int argc,char *argv[])
 
 	return 0;
 }
-
-#endif
 
 #endif
