+++ bspatch.c
@@ -25,17 +25,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <stdint.h>
-
-struct bspatch_stream
-{
-	void* opaque;
-	int (*read)(const struct bspatch_stream* stream, void* buffer, int length);
-};
-
-int bspatch(const uint8_t* old, int64_t oldsize, uint8_t* new, int64_t newsize, struct bspatch_stream* stream);
-
-#if !defined(BSPATCH_HEADER_ONLY)
+#include "bspatch.h"
 
 static int64_t offtin(uint8_t *buf)
 {
@@ -194,5 +184,4 @@ int main(int argc,char * argv[])
 	return 0;
 }
 
-#endif
 #endif
